Channel: remove unimplemented close method
The original plan was to have close(), matching channel state
"closed" and Closable interface. The expectation of closing
channels in a blocking manner as it would be expected by close()
did not seem like a good idea, though. Instead terminateAsync()
should be used, possibly followed by timed waiting, when actually
needed. close() is probably not what most people want, at least not
by what it may imply, so it should better be removed to avoid
confusion.